Verificar se os arquivos SIGTAP existem

+++ R/4_2_get_counties_CNES.R
@@ -28,8 +28,22 @@ get_counties <- function(){
     dplyr::select(-municipios)
 
 
-  #Realizando a primeira chamada da função download_cnes_files
-  download_cnes_files(newer = TRUE)
+
+  caminho = here::here("data-raw/CNES/ST")
+  if (dir.exists(caminho)) {
+    dbc_files <- list.files(caminho, pattern = "\\.dbc$", full.names = TRUE)
+    if (length(dbc_files) == 27){
+      cat("Os arquivos ST(Servicos Temporarios) do CNES nao serao baixados, pois ja estao no diretorio ./data-raw/CNES/ST.\n")
+    } else if(length(dbc_files)>=1) {
+      stop("A pasta ./data-raw/CNES/ST nao contem todos os 27 arquivos DBC do ST(Servicos Temporarios) do CNES")
+    } else {
+      stop("A pasta ./data-raw/CNES/ST nao contem nenhum dos 27 arquivos DBC do ST(Servicos Temporarios) do CNES")
+    }
+
+  } else {
+    #Realizando a primeira chamada da função download_cnes_files
+    download_cnes_files(newer = TRUE)
+  }
 
   health_establishment <-
     here::here("data-raw", "CNES", "ST") %>%
@@ -37,30 +51,39 @@ get_counties <- function(){
     purrr::map_dfr(read.dbc::read.dbc, as.is=TRUE)
 
 
-  #Movendo os arquivos do CADGER para a pasta "data-raw/CNES/CADGER"
-  #------------------------------------------------------------------------
-  output_dir <- here::here("data-raw", "CNES", "CADGER")
-  dir.create(output_dir)
+  caminho2 = here::here("data-raw/CNES/CADGER")
+  if (dir.exists(caminho2)) {
+    dbf_files <- list.files(caminho2, pattern = "\\.dbf$", full.names = TRUE)
+    if (length(dbf_files) == 28){
+      cat("Os arquivos CADGER do CNES nao serao baixados, pois ja estao no diretorio ./data-raw/CNES/CADGER.\n")
+    } else if(length(dbf_files)>=1) {
+      stop("A pasta ./data-raw/CNES/CADGER nao contem todos os 28 arquivos DBF do CADGER do CNES")
+    } else {
+      stop("A pasta ./data-raw/CNES/CADGER nao contem nenhum dos 28 arquivos DBF do CADGER do CNES")
+    }
+  } else {
+    #Movendo os arquivos do CADGER para a pasta "data-raw/CNES/CADGER"
+    output_dir <- here::here("data-raw", "CNES", "CADGER")
+    dir.create(output_dir)
 
-  caminho_pasta <- system.file("extdata", package = "DATASUS.SIA.SIH")
-  caminho_completo <- file.path(caminho_pasta, "CADGER.zip")
-  dir_destino <- "./data-raw/CNES/CADGER"
+    caminho_pasta <- system.file("extdata", package = "DATASUS.SIA.SIH")
+    caminho_completo <- file.path(caminho_pasta, "CADGER.zip")
+    dir_destino <- "./data-raw/CNES/CADGER"
 
-  # Verifique se o arquivo existe
-  if (file.exists(caminho_completo)) {
-    novo_caminho_completo <- file.path(dir_destino, "CADGER.zip")
-    file.copy(caminho_completo, novo_caminho_completo)
-    #cat("Arquivos CADGER movido com sucesso para:", novo_caminho_completo, "\n")
-  } else {cat("O arquivo 'CADGER.zip' nao foi encontrado no diretório:", caminho_pasta, "\n")}
+    # Verifique se o arquivo existe
+    if (file.exists(caminho_completo)) {
+      novo_caminho_completo <- file.path(dir_destino, "CADGER.zip")
+      file.copy(caminho_completo, novo_caminho_completo)
+      #cat("Arquivos CADGER movido com sucesso para:", novo_caminho_completo, "\n")
+    } else {cat("O arquivo 'CADGER.zip' nao foi encontrado no diretório:", caminho_pasta, "\n")}
 
-  #Extraindo arquivo CADGER.zip
-  unzip('./data-raw/CNES/CADGER/CADGER.zip', exdir = './data-raw/CNES/CADGER')
+    #Extraindo arquivo CADGER.zip
+    unzip('./data-raw/CNES/CADGER/CADGER.zip', exdir = './data-raw/CNES/CADGER')
 
-  if (file.exists("./data-raw/CNES/CADGER/CADGER.zip")) {
-    file.remove('./data-raw/CNES/CADGER/CADGER.zip')
+    if (file.exists("./data-raw/CNES/CADGER/CADGER.zip")) {
+      file.remove('./data-raw/CNES/CADGER/CADGER.zip')
+    }
   }
-  #------------------------------------------------------------------------
-
   #Le todos os arquivos DBC do diretorio "data-raw/CNES/CADGER" e combina-os em uma tabela unica.
   health_establishment_details <-
     here::here("data-raw", "CNES", "CADGER") %>%
